+++ src/index.ts
@@ -41,3 +41,4 @@ export * from './components/spinner';
 export * from './components/suggest';
 export * from './components/highlight';
 export * from './components/suggest-input';
+export * from './components/text-label';